+++ lib/td_dd/canada/data_structure_type_abilities.ex
@@ -3,6 +3,9 @@ defmodule TdDd.Canada.DataStructureTypeAbilities do
   alias TdDd.Auth.Claims
   alias TdDd.DataStructures.DataStructureType
 
+  # Admin accounts can do anything with data structure types
+  def can?(%Claims{role: "admin"}, _action, _resource), do: true
+
   def can?(%Claims{}, :index, %DataStructureType{}), do: true
 
   def can?(%Claims{}, :index, DataStructureType), do: true