+++ test/td_cx_web/controllers/job_controller_test.exs
@@ -149,14 +149,4 @@ defmodule TdCxWeb.JobControllerTest do
     job = insert(:job)
     {:ok, job: job}
   end
-
-  defp create_acl_entry(user_id, domain_id, permissions) do
-    MockPermissionResolver.create_acl_entry(%{
-      principal_id: user_id,
-      principal_type: "user",
-      resource_id: domain_id,
-      resource_type: "domain",
-      permissions: permissions
-    })
-  end
 end