+++ CHANGELOG.md
@@ -7,6 +7,7 @@
 ### Fixed
 
 - [TD-3235] Fallback of uncontroller responses on metadata controller
+
 ## [4.15.0] 2021-03-08
 
 ### Changed