+++ CHANGELOG.md
@@ -2,7 +2,13 @@
 
 ## [Unreleased] 
 
+### Added
+
 - [TD-2951] `profile_structure` permission
+
+### Removed
+
+- [TD-3421] remove `/data_structures/search/source_alias` endpoint
 
 ## [4.15.0] 2021-03-08