+++ CHANGELOG.md
@@ -6,6 +6,10 @@
 
 - [TD-2951] `profile_structure` permission
 
+### Fixed
+
+- [TD-3235] Fallback of uncontroller responses on metadata controller
+
 ### Removed
 
 - [TD-3421] remove `/data_structures/search/source_alias` endpoint